pyproject.toml
[project]
name = "dredge"
version = "0.1.0"
description = "Cloud Incident Response & Threat Hunting for AWS + GitHub"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name="Solidarity Labs", email="sabastante@solidaritylabs.io" }
]

dependencies = [
    "boto3>=1.33.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "urllib3>=2.0.0",
    "google-cloud-logging>=3.10",
    "google-auth>=2.0",
]

[project.scripts]
dredge = "dredge.cli:main"

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

